test of tests, refactor of game.js, added gameView.js, seperating state from view
