+++ infection/src/components/game.js
@@ -1,16 +1,8 @@
 import React, { Component} from 'react';
-import { Grid, Row, Col } from 'react-bootstrap';
 
 import socket from '../socket';
 
-import Roles from '../views/game/roles/roles';
-import Round from '../views/game/round/round';
-import Mission from '../views/game/mission/mission';
-import MissionResults from '../views/game/missionResults/missionResults';
-import WaitingForTeam from '../views/game/waiting/waitingForTeam';
-import GameOver from '../views/game/gameOver/gameOver';
-import GameStatus from '../views/game/gameStatus/gameStatus';
-import Header from '../views/game/shared/header';
+import GameView from '../views/game/gameView';
 
 class Game extends Component {
   constructor(props) {
@@ -96,48 +88,14 @@ class Game extends Component {
   }
   
   render() {
-    const game = this.state;
-
-    return <Grid className="game">
-      <br></br>
-      <Header></Header>
-      <br></br>
-      <Row>
-        <Col med={2}></Col>
-        <Col med={8}>
-          {
-            !game.teamAssembled
-              ? <WaitingForTeam></WaitingForTeam>
-              : !game.round
-                ? <Roles infiltrator={game.infiltrator}></Roles>
-                : !game.missionActive
-                  ? <Round
-                    game={game}
-                    handleSelectRosterEntryClick={this.handleSelectRosterEntryClick.bind(this)}
-                    handleSubmitRoster={this.handleSubmitRoster.bind(this)}
-                  ></Round>
-                  : !game.missionResults[game.round - 1]
-                    ? <Mission
-                      choose={this.handleOnMissionClick.bind(this)}
-                      choiceMade={game.choiceMade}
-                      roster={game.missionRoster}
-                      username={game.username}
-                    ></Mission>
-                    : !game.gameOver
-                      ? <MissionResults
-                        result={game.missionResults[game.round - 1]}
-                      ></MissionResults>
-                      : <GameOver scientistsWin={game.scientistsWin}></GameOver>
-          }
-        </Col>
-        <Col med={2}></Col>
-      </Row>
-      <Row>
-        <br></br>
-        <br></br>
-      </Row>
-          <GameStatus missionResults={game.missionResults}></GameStatus>
-    </Grid>
+    return (
+      <GameView 
+        game={this.state}
+        handleSelectRosterEntryClick={this.handleSelectRosterEntryClick.bind(this)}
+        handleSubmitRoster={this.handleSubmitRoster.bind(this)}
+        choose={this.handleOnMissionClick.bind(this)}
+      ></GameView>
+    );
   }
 }
